requirements.txt
python-telegram-bot==10.1.0
iexfinance==0.3.4
